requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
dvc[s3]
mlflow==2.2.2
pytest
pandera
fastapi
uvicorn
python-multipart
joblib
pymongo
python-dotenv
ipykernel
ruff
loguru
-e .